Update KSP to 20201110
This PR updates KSP version to 20201110.

Unfortunately, KSP changed how it works where it no longer participates
in the main kotlin compilation, instead, it runs as a dependency task
to the main compilation where it runs the analysis and adds the new source
to the main compilation.

This PR reverts that behavior by overriding the extension's analysis results.
It is a quick fix rather than a proper one. For the proper fix, we either
need to figure out how to run KSP compilation separately (which would require
changing the main project) or have a different integration for KSP where we
ask the user to do it (which seems not good for usability).